+++ lua/pack.lua
@@ -222,6 +222,7 @@ conform.setup({
         go               = { "goimports", "gofmt" },
         sh               = { "shfmt" },
         bash             = { "shfmt" },
+        zsh              = { "shfmt" },
         terraform        = { "terraform_fmt" },
         tf               = { "terraform_fmt" },
         ["terraform-vars"] = { "terraform_fmt" },